requirements.txt
astroid==2.9.2
attrs==21.2.0
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
charset-normalizer==2.0.10
distlib==0.3.4
filelock==3.4.0
idna==3.3
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.7.1
lxml==4.6.4
mccabe==0.6.1
packaging==21.3
Pillow==8.4.0
pipenv==2021.11.23
platformdirs==2.4.0
pluggy==1.0.0
pptx2md==1.0.0
py==1.11.0
pylint==2.12.2
pyparsing==3.0.6
pytest==6.2.5
python-pptx==0.6.21
PyYAML==6.0
rapidfuzz==1.9.0
requests==2.27.0
six==1.16.0
toml==0.10.2
typing_extensions==4.0.1
unotools==0.3.3
urllib3==1.26.7
virtualenv==20.10.0
virtualenv-clone==0.5.7
wrapt==1.13.3
XlsxWriter==3.0.2
